refactor: common headers decomposition
